Simplify component by refactoring out a TemplateSelectorItem component
This limits the noise that was buiilding in the TemplateSelectorControl component file.
